requirements.txt
streamlit
sqlalchemy
langchain
langchain-community
langchain-groq
duckduckgo-search
python-dotenv

